% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis.R
\name{DE.analysis}
\alias{DE.analysis}
\title{Automated Differential Expression Pipeline for RNA-Seq Counts using DESeq2 or edgeR}
\arguments{
\item{M}{Character. Choose method, either "DESeq2" or "edgeR".}

\item{IN}{A numeric matrix or data frame of raw RNA-Seq read counts (file path).}

\item{S}{Numeric pair. The number of samples in the dataset (x:y).}

\item{OUT}{Character. Prefix for the exported text and PDF files.}

\item{LF}{Numeric. Log fold change threshold}

\item{FD}{Numeric. adjusted p-value threshold.}
}
\description{
Automated Differential Expression Pipeline for RNA-Seq Counts using DESeq2 or edgeR
}
